Separate tasks into two files
